+++ CopilotKit/examples/next-openai/CHANGELOG.md
@@ -1,5 +1,17 @@
 # web
 
+## 1.4.2-tyler-reset-chat.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-core@1.5.0-tyler-reset-chat.0
+  - @copilotkit/react-ui@1.5.0-tyler-reset-chat.0
+  - @copilotkit/react-textarea@1.5.0-tyler-reset-chat.0
+  - @copilotkit/runtime@1.5.0-tyler-reset-chat.0
+  - @copilotkit/runtime-client-gql@1.5.0-tyler-reset-chat.0
+  - @copilotkit/shared@1.5.0-tyler-reset-chat.0
+
 ## 1.4.1
 
 ### Patch Changes