+++ CopilotKit/packages/react-textarea/CHANGELOG.md
@@ -1,5 +1,18 @@
 # ui
 
+## 1.5.0-tyler-reset-chat.0
+
+### Minor Changes
+
+- Merging Markus' reset fixes with main + a new change for stopping an agent during execution of FE actions
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-core@1.5.0-tyler-reset-chat.0
+  - @copilotkit/runtime-client-gql@1.5.0-tyler-reset-chat.0
+  - @copilotkit/shared@1.5.0-tyler-reset-chat.0
+
 ## 1.4.1
 
 ### Patch Changes